Almost there

+++ server/src/main/java/projeto/server/TaskRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.web.bind.annotation.CrossOrigin;
 
 @CrossOrigin
-public interface TaskRepository extends JpaRepository<Task, Long> { }
+public interface TaskRepository extends JpaRepository<Task, Integer> { }
